RAG/RAG.requirements.txt
beautifulsoup4==4.13.3
fastapi==0.115.11
FlagEmbedding==1.3.4
langchain==0.3.19
langchain_core==0.3.40
langchain_qdrant==0.2.0
langgraph==0.3.5
numpy==2.2.3
openai-whisper
Pillow==11.1.0
protobuf==5.29.3
python-dotenv==1.0.1
qdrant_client==1.13.2
RAG==0.1.0
Requests==2.32.3
uvicorn==0.34.0